Change class visibility and move Program classes to DLLInterop.
